requirements.txt
#pydensecrf - install manually
git+https://github.com/lucasb-eyer/pydensecrf.git
matplotlib
numpy
scipy==1.2.0
torch
torchvision==0.2.0
tqdm
protobuf
tensorboardX
opencv-python
scikit-image
Pillow
